Moved overflow higher up and deleted the component
